% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{applyItersplits}
\alias{applyItersplits}
\title{applyItersplits}
\usage{
applyItersplits(iters, splits, replace = FALSE)
}
\arguments{
\item{iters}{number of iterations}

\item{splits}{list of vectors of row numbers}
}
\description{
generate splits for splithalf
}
