% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colMedians}
\alias{colMedians}
\alias{colMedians_mask}
\alias{mediansByMask}
\alias{colMeans_mask}
\alias{meansByMask}
\title{colMedians}
\usage{
colMedians(mat)

colMedians_mask(mat, mask)

mediansByMask(values, mask)

colMeans_mask(mat, mask)

meansByMask(values, mask)
}
\arguments{
\item{mat}{a matrix with values to aggregate}

\item{mask}{a logical matrix determining which data points to include and which not to}

\item{values}{Values to aggregate over in different mask configurations}
}
\description{
get column medians
}
