% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colSds}
\alias{colSds}
\alias{colSds_mask}
\alias{sdsByMask}
\title{colSds}
\usage{
colSds(mat)

colSds_mask(mat, mask)

sdsByMask(values, mask)
}
\arguments{
\item{mat}{the matrix to retrieve column SDs from}

\item{mask}{a logical matrix determining which data points to include and which not to}

\item{values}{Values to aggregate over in different mask configurations}
}
\description{
get column SDs
}
