% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormean.R
\name{comprel}
\alias{comprel}
\title{Compare split-half reliabilities}
\usage{
comprel(x, y, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{x, y}{Either \code{rapidsplit} objects outputted by \code{\link[=rapidsplithalf]{rapidsplithalf()}}, or
raw vectors or permutation-based split-half reliabilities.}

\item{alternative}{The type of test to perform: "two.sided",
"less" (x < y), or "greater" (x > y).}
}
\value{
The p-value for the difference.
}
\description{
Compare split-half reliabilities
}
\details{
For each split-half correlation in \code{x},
this computes the percentage of individual split-half correlations in \code{y}
that are smaller; these percentages are then averaged and modified in accordance
with the test requested in \code{alternative} to give the p-value.
}
\note{
A large number of splits (>10,000) is recommended to get an accurate p-value.
}
\examples{

# Here we will demonstrate that the reliability of the practice trials in the IAT
# is higher than the reliability of the non-practice trials.
data(raceIAT)
rel1 <- rapidsplit(data=raceIAT[raceIAT$blocktype=="practice",],
                   subjvar="session_id",diffvars="congruent",
                   subscorevar="blocktype",aggvar="latency",splits=1000,standardize=TRUE)
                   
rel2 <- rapidsplit(data=raceIAT[raceIAT$blocktype=="full",],
                   subjvar="session_id",diffvars="congruent",
                   subscorevar="blocktype",aggvar="latency",splits=1000,standardize=TRUE)
comprel(rel1,rel2,alternative="greater")

}
\author{
Sercan Kahveci
}
