% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ExcludeSDOutliers}
\alias{ExcludeSDOutliers}
\alias{ExcludeSDOutliers_nomask}
\title{Exclude SD-based outliers}
\usage{
ExcludeSDOutliers(rtvec, mask, sdlim = 3)

ExcludeSDOutliers_nomask(mat, sdlim = 3)
}
\arguments{
\item{rtvec}{Reaction time vector}

\item{mask}{a logical matrix determining which data points to include and which not to}

\item{sdlim}{Standard deviation limit to apply; values beyond are classified as outliers and masked}

\item{mat}{Matrix in which to mark SD-based outleirs by column (with FALSE)}
}
\value{
An updated mask
}
\description{
Update a mask matrix based on outlyingness
}
