% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maskOLExclusion}
\alias{maskOLExclusion}
\title{Exclude outliers by mask}
\usage{
maskOLExclusion(values, mask, maxsd)
}
\arguments{
\item{values}{values to detect outliers from}

\item{mask}{a logical matrix determining which data points to include per iteration (column)}

\item{maxsd}{values exceeding more than this value in SDs will be excluded from the mask}
}
\value{
An updated logical mask matrix with all outliers excluded
}
\description{
Exclude outliers by mask
}
