% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{rapidsplithalf}
\alias{rapidsplithalf}
\alias{rapidsplithalf-package}
\title{rapidsplithalf package}
\description{
Accurately estimates the reliability of cognitive tasks using a fast and flexible 
permutation-based split-half reliability algorithm 
that supports stratified splitting while maintaining equal split sizes. 
See Kahveci, Bathke, and Blechert (2025) \doi{10.3758/s13423-024-02597-y} for details.
To learn more about rapidsplithalf, view the introductory vignette:
\code{vignette("rapidsplithalf",package="rapidsplithalf")}
}
\author{
Sercan Kahveci
}
