% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spearmanbrown.R
\name{requiredTestSize}
\alias{requiredTestSize}
\title{Required test size for target reliability}
\usage{
requiredTestSize(
  obs_r,
  goal_r = 0.8,
  n = 1,
  fix.negative = c("mirror", "nullify", "none")
)
}
\arguments{
\item{obs_r}{The observed reliability}

\item{goal_r}{The goal reliability}

\item{n}{Current test size, i.e., the number of items or trials.}

\item{fix.negative}{How should negative observed reliabilities be handled? See \code{\link[=spearmanBrown]{spearmanBrown()}}.}
}
\value{
The required test size.
}
\description{
Compute how many items a test needs, for its reliability to reach a goal value.
}
\details{
This function is obtained by solving the \code{\link[=spearmanBrown]{spearmanBrown()}} formula for the \code{ntests} argument.
}
\examples{
requiredTestSize(obs_r=0.5, goal_r=0.8,n=256)

requiredTestSize(obs_r=0, goal_r=0.8,n=256)

}
\author{
Sercan Kahveci
}
