% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corByColumns}
\alias{corByColumns}
\alias{corByColumns_mask}
\title{Correlate each column of 1 matrix with the same column in another matrix}
\usage{
corByColumns(mat1, mat2)

corByColumns_mask(mat1, mat2, mask)
}
\arguments{
\item{mat1, mat2}{Matrices to correlate}

\item{mask}{Logical matrix marking which data points to include}
}
\value{
A numeric vector of correlations per column
}
\description{
Correlate each column of 1 matrix with the same column in another matrix
}
