% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapidsplit.R, R/zzz.R
\name{rapidsplit}
\alias{rapidsplit}
\alias{rapidsplit-package}
\title{rapidsplit}
\usage{
rapidsplit(
  ds,
  subjvar,
  diffvars = NULL,
  stratvars = NULL,
  rtvar,
  iters,
  agg = c("means", "medians"),
  standardize = F
)
}
\arguments{
\item{ds}{Dataset, a \code{data.frame}}

\item{subjvar}{Subject ID variable name, a \code{character}}

\item{diffvars}{Variables that determine which conditions 
need to be subtracted from each other, a \code{character}}

\item{stratvars}{Additional variables that the splits should 
be stratified by, if possible; a \code{character}}

\item{rtvar}{Reaction time variable name, a \code{character}}

\item{iters}{Number of split-halves to average, an \code{integer}}

\item{agg}{The function by which to aggregate the RTs; can be \code{"means"} or \code{"medians"}}

\item{standardize}{Whether to divide by scores by the subject's SD; a \code{logical}}
}
\value{
A \code{list} containing the averaged reliability as well as 
a vector with the reliability of each iteration
}
\description{
A very fast algorithm for permutated split-half reliability
}
\examples{
print("no example")
}
