% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapidsplit.R, R/zzz.R
\docType{package}
\name{rapidsplit}
\alias{rapidsplit}
\title{rapidsplit
ultra fast split-half}
\usage{
rapidsplit(
  ds,
  subjvar,
  pullvar,
  targetvar,
  rtvar,
  iters,
  agg = c("means", "medians"),
  standardize = F
)
}
\arguments{
\item{ds}{dataset}

\item{subjvar}{subject var name}

\item{pullvar}{movement direction var name}

\item{targetvar}{stim type var name}

\item{rtvar}{rt varname}

\item{iters}{n iterations}

\item{agg}{means or medians}

\item{standardize}{divide by sd or not}
}
\value{

}
\examples{
print("no example")
}
